rodio-node/requirements.txt
web3==6.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
aiohttp==3.9.0
paho-mqtt==1.6.1
python-multipart==0.0.6
httpx==0.25.0